third_party/blink/web_tests/svg/parser/whitespace-length-invalid-4-expected.txt

CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ". ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".E0%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "e1%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " . ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .E0%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " e1%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   . ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .E0%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   e1%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t . ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .E0%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t e1%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1 ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.E0%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\fe1%\f".
This is a testharness.js-based test.
Harness: the test ran to completion.
